requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
cloudinary==1.24.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-easy-select2==1.5.7
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
packaging==20.9
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
zope.interface==5.4.0
